requirements.txt
kafka-python-ng
crc32c
iris-pex-embedded-python>=3.0.0
pytest
sqlalchemy-iris